Dedup models commands tests
